settings: allow to disable quicksettings
on 4.3 we had an extra settings option for it. Due that we want
to reduce the settings to improve the UX and due that
we are doing it this way on all dslv settings we change
the old 4.3 way.

Now it is very easy like dslv.....remove all tiles and QS
is completly disabled and cannot be opend any more. Notification
drawer shows the settings icon on the top right in this case.
As well compared to 4.3 this works on phones and tablets

Code cleanup and follow AOSP code style guidelines.

Some minor dialog tweeks (help dialog etc)

Change-Id: Ic3007e3cff0a093b29c28f32ffc0d0951a771b93

Conflicts:
	res/xml/notification_drawer_qs_settings.xml
	src/com/android/settings/liquid/quicksettings/QuickSettingsTiles.java