Settings: Super advanced Custom Toggle (2/2)
If user sets up a custom tile with Chamber of Secrets
actions all of the same setting and each action only
has one value per tile-state we now set up an observer
for that setting. This means the tile will listen to
settings changes for the user and change states as
necessary.

EXAMPLE:
Custom Tile Settings:
State 1 > Chamber > "location_mode" > 3 > name: "GPS on"
State 2 > Chamber > "location_mode" > 2 > name: "GPS off"
Save.

User is informed their tile can now be observed and will change
states as necessary.  Ask them to add an icon and label if the
state isn't recognized (I.E. location_mode is set to 1).
Label set to "disabled" with [ICON].
GPS is changed via Settings>Location...
Another tile...etc.
Tile matches it's state or moves to unknown if setting != user
values in tile.

See Video:
https://www.youtube.com/watch?v=pw5Epcp0fq4

PSX:
Reverted to old build and dirty flashed.
Fixed potential FC's there.  This code can
probably stay since it's all Settings-side...
But if we'd like we can phase it out in a
couple months or something.  We don't want
SystemUI to take a dump when a user saves a tile
that has old code in it's settings.

Change-Id: I09034c76c3bc67179c6eaeb0b23a6d39415caf68

Conflicts:
	src/com/android/settings/liquid/quicksettings/QuickSettingsTiles.java